moved data loading around, advanced highlevel interface to care2d
